requirements.txt
# Core
python-dotenv
Flask

# Machine Learning
scikit-learn
pandas
joblib

# OpenAI (for GPT integration)
openai

# Optional (data visualization / debugging)
matplotlib==3.9.2
seaborn==0.13.2

# Optional for Gimini integration
requests==2.31.0
httpx==0.24.1
tenacity==8.2.2